Moving code to deletes_and_updates folder.
